+++ packages/swag-axios-js/CHANGELOG.md
@@ -1,5 +1,15 @@
 # apipgen-swag-js
 
+## 0.1.9
+
+### Patch Changes
+
+- chore: fix transform return type error
+- Updated dependencies
+  - @apipgen/swag-parser@0.1.8
+  - @apipgen/pipeline@0.1.5
+  - apipgen@0.3.10
+
 ## 0.1.8
 
 ### Patch Changes